+++ packages/components/src/Checkbox/CheckboxGroup.tsx
@@ -80,7 +80,7 @@ export const CheckboxGroup = ({
   return (
     <FieldBase
       label={props.label}
-      labelProps={{ as: 'span', ...labelProps }}
+      labelProps={{ elementType: 'span', ...labelProps }}
       description={props.description}
       descriptionProps={descriptionProps}
       error={error}