+++ packages/components/src/Checkbox/_Checkbox.test.tsx
@@ -6,7 +6,7 @@ import { Theme, cva } from '@marigold/system';
 
 import { FieldGroup } from '../FieldBase';
 import { setup } from '../test.utils';
-import { Checkbox } from './Checkbox';
+import { Checkbox } from './_Checkbox';
 
 const theme: Theme = {
   name: 'test',
@@ -47,12 +47,12 @@ const { render } = setup({ theme });
 // Tests
 // ---------------
 test('renders label and (hidden) checkbox', () => {
-  render(<Checkbox data-testid="checkboxId">With Label</Checkbox>);
+  render(<Checkbox data-testid="checkbox">With Label</Checkbox>);
 
   const label = screen.getByText('With Label');
   expect(label).toBeInTheDocument();
 
-  const checkbox = screen.getAllByTestId('checkboxId')[1];
+  const checkbox = screen.getByTestId('checkbox');
   expect(checkbox).toBeInTheDocument();
   expect(checkbox).toBeInstanceOf(HTMLInputElement);
   expect(checkbox).toHaveAttribute('type', 'checkbox');
@@ -61,7 +61,7 @@ test('renders label and (hidden) checkbox', () => {
 test('allows to render without label', () => {
   render(<Checkbox data-testid="checkbox" aria-label="No Label" />);
 
-  const checkbox = screen.getAllByTestId('checkbox')[1];
+  const checkbox = screen.getByTestId('checkbox');
   expect(checkbox).toBeInTheDocument();
   expect(checkbox).toBeInstanceOf(HTMLInputElement);
   expect(checkbox).toHaveAttribute('type', 'checkbox');
@@ -76,7 +76,7 @@ test('supports read only state', () => {
     </Checkbox>
   );
 
-  const checkbox = screen.getAllByTestId<HTMLInputElement>('checkbox')[1];
+  const checkbox = screen.getByTestId<HTMLInputElement>('checkbox');
   const component = screen.getByText('Read Only');
 
   fireEvent.click(component);
@@ -89,9 +89,13 @@ test('check if all slot class names are applied correctly', () => {
   const label = screen.getByText('With Label');
 
   expect(label.parentElement?.className).toMatchInlineSnapshot(
-    `"group/checkbox flex items-center gap-[0.5rem] cursor-pointer data-[disabled]:cursor-not-allowed"`
+    `"group/checkbox relative flex items-center gap-[1ch]"`
   );
-
+  expect(
+    (label.parentElement?.childNodes[0] as HTMLElement).className
+  ).toMatchInlineSnapshot(
+    `"z-1 absolute left-0 top-0 h-full w-full cursor-pointer opacity-[0.0001] group-disabled/checkbox:cursor-not-allowed"`
+  );
   expect(getVisibleCheckbox()?.className).toMatchInlineSnapshot(
     `"flex shrink-0 grow-0 basis-4 items-center justify-center h-4 w-4 border border-solid border-checkbox-base-border bg-checkbox-base-background rounded-[2] p-0.5 data-[hover]:border-checkbox-base-hover data-[focus]:outline-offset[3] data-[focus]:outline data-[focus]:outline-2 data-[checked]:border-checkbox-base-checked data-[checked]:bg-checkbox-base-checkedBackground data-[checked]:text-white data-[indeterminate]:border-checkbox-base-indeterminate data-[indeterminate]:bg-checkbox-base-indeterminateBackground data-[indeterminate]:text-white data-[disabled]:border-checkbox-base-disabled data-[disabled]:bg-checkbox-base-disabledBackground"`
   );
@@ -120,7 +124,7 @@ test('correct class name is set on size small', () => {
   const label = screen.getByText('With Label');
 
   expect(label.parentElement?.className).toMatchInlineSnapshot(
-    `"group/checkbox flex items-center gap-[0.5rem] cursor-pointer data-[disabled]:cursor-not-allowed py-1"`
+    `"group/checkbox relative flex items-center gap-[1ch] py-1"`
   );
 });
 
@@ -131,9 +135,7 @@ test('support default checked', () => {
     </Checkbox>
   );
 
-  const input: HTMLInputElement = screen.getAllByTestId(
-    'checkbox'
-  )[1] as HTMLInputElement;
+  const input: HTMLInputElement = screen.getByTestId('checkbox');
   expect(input.checked).toBeTruthy();
 });
 
@@ -143,9 +145,7 @@ test('supports indeterminate state', () => {
       With Label
     </Checkbox>
   );
-  const input: HTMLInputElement = screen.getAllByTestId(
-    'checkbox'
-  )[1] as HTMLInputElement;
+  const input: HTMLInputElement = screen.getByTestId('checkbox');
   expect(input.indeterminate).toBeTruthy();
 });
 
@@ -156,9 +156,7 @@ test('controlled', () => {
       With Label
     </Checkbox>
   );
-  const input: HTMLInputElement = screen.getAllByTestId(
-    'checkbox'
-  )[1] as HTMLInputElement;
+  const input: HTMLInputElement = screen.getByTestId('checkbox');
 
   fireEvent.click(input);
   expect(onChange).toHaveBeenCalledWith(true);
@@ -185,6 +183,6 @@ test('works with a <FieldGroup>', () => {
     </FieldGroup>
   );
 
-  const checkbox = screen.getAllByTestId<HTMLInputElement>('checkbox')[1];
+  const checkbox = screen.getByTestId('checkbox');
   expect(checkbox).toBeInTheDocument();
 });